refactor: add pausable and rename some symbols
