+++ src/config/appDataTypes.ts
@@ -3,7 +3,7 @@ import { AppDataRecord } from '@graasp/sdk/frontend';
 
 import { List } from 'immutable';
 
-import { ProcessState } from '@/interfaces/interactionProcess';
+import { ActivityType, ProcessStatus } from '@/interfaces/interactionProcess';
 import { NoveltyRelevanceRatings } from '@/interfaces/ratings';
 
 export enum AppDataTypes {
@@ -43,7 +43,8 @@ export type IdeaSetAppData = AppDataRecord & {
 
 export type CurrentStateData = {
   round?: number;
-  state: ProcessState;
+  status: ProcessStatus;
+  activity: ActivityType;
 };
 
 export type CurrentStateAppData = AppDataRecord & {
